+++ tket/src/Circuit/CircPool.cpp
@@ -17,8 +17,10 @@
 #include <tkassert/Assert.hpp>
 
 #include "Circuit.hpp"
+#include "Gate/Rotation.hpp"
 #include "OpType/OpType.hpp"
 #include "Utils/Expression.hpp"
+#include "Utils/MatrixAnalysis.hpp"
 
 namespace tket {